Move viewport functions into ui
