+++ src/openrct2-ui/ride/Construction.cpp
@@ -104,4 +104,133 @@ namespace OpenRCT2
         }
         return list;
     }
+
+    /**
+     *
+     *  rct2: 0x006CCF70
+     */
+    CoordsXYZD RideGetEntranceOrExitPositionFromScreenPosition(const ScreenCoordsXY& screenCoords)
+    {
+        CoordsXYZD entranceExitCoords{};
+        gRideEntranceExitPlaceDirection = INVALID_DIRECTION;
+        // determine if the mouse is hovering over a station - that's the station to add the entrance to
+        auto info = GetMapCoordinatesFromPos(screenCoords, EnumsToFlags(ViewportInteractionItem::Ride));
+        if (info.SpriteType != ViewportInteractionItem::None)
+        {
+            if (info.Element->GetType() == TileElementType::Track)
+            {
+                const auto* trackElement = info.Element->AsTrack();
+                if (trackElement->GetRideIndex() == gRideEntranceExitPlaceRideIndex)
+                {
+                    const auto& ted = GetTrackElementDescriptor(trackElement->GetTrackType());
+                    if (std::get<0>(ted.SequenceProperties) & TRACK_SEQUENCE_FLAG_ORIGIN)
+                    {
+                        if (trackElement->GetTrackType() == TrackElemType::Maze)
+                        {
+                            gRideEntranceExitPlaceStationIndex = StationIndex::FromUnderlying(0);
+                        }
+                        else
+                        {
+                            gRideEntranceExitPlaceStationIndex = trackElement->GetStationIndex();
+                        }
+                    }
+                }
+            }
+        }
+
+        auto ride = GetRide(gRideEntranceExitPlaceRideIndex);
+        if (ride == nullptr)
+        {
+            entranceExitCoords.SetNull();
+            return entranceExitCoords;
+        }
+
+        auto stationBaseZ = ride->GetStation(gRideEntranceExitPlaceStationIndex).GetBaseZ();
+
+        auto coordsAtHeight = ScreenGetMapXYWithZ(screenCoords, stationBaseZ);
+        if (!coordsAtHeight.has_value())
+        {
+            entranceExitCoords.SetNull();
+            return entranceExitCoords;
+        }
+
+        entranceExitCoords = { coordsAtHeight->ToTileStart(), stationBaseZ, INVALID_DIRECTION };
+
+        if (ride->type == RIDE_TYPE_NULL)
+        {
+            entranceExitCoords.SetNull();
+            return entranceExitCoords;
+        }
+
+        auto stationStart = ride->GetStation(gRideEntranceExitPlaceStationIndex).Start;
+        if (stationStart.IsNull())
+        {
+            entranceExitCoords.SetNull();
+            return entranceExitCoords;
+        }
+
+        // find the quadrant the mouse is hovering over - that's the direction to start searching for a station TileElement
+        Direction startDirection = 0;
+        auto mapX = (coordsAtHeight->x & 0x1F) - 16;
+        auto mapY = (coordsAtHeight->y & 0x1F) - 16;
+        if (std::abs(mapX) < std::abs(mapY))
+        {
+            startDirection = mapY < 0 ? 3 : 1;
+        }
+        else
+        {
+            startDirection = mapX < 0 ? 0 : 2;
+        }
+        // check all 4 directions, starting with the mouse's quadrant
+        for (uint8_t directionIncrement = 0; directionIncrement < 4; directionIncrement++)
+        {
+            entranceExitCoords.direction = (startDirection + directionIncrement) & 3;
+            // search for TrackElement one tile over, shifted in the search direction
+            auto nextLocation = entranceExitCoords;
+            nextLocation += CoordsDirectionDelta[entranceExitCoords.direction];
+            if (MapIsLocationValid(nextLocation))
+            {
+                // iterate over every element in the tile until we find what we want
+                auto* tileElement = MapGetFirstElementAt(nextLocation);
+                if (tileElement == nullptr)
+                    continue;
+                do
+                {
+                    if (tileElement->GetType() != TileElementType::Track)
+                        continue;
+                    if (tileElement->GetBaseZ() != stationBaseZ)
+                        continue;
+                    auto* trackElement = tileElement->AsTrack();
+                    if (trackElement->GetRideIndex() != gRideEntranceExitPlaceRideIndex)
+                        continue;
+                    if (trackElement->GetTrackType() == TrackElemType::Maze)
+                    {
+                        // if it's a maze, it can place the entrance and exit immediately
+                        entranceExitCoords.direction = DirectionReverse(entranceExitCoords.direction);
+                        gRideEntranceExitPlaceDirection = entranceExitCoords.direction;
+                        return entranceExitCoords;
+                    }
+                    // if it's not a maze, the sequence properties for the TrackElement must be found to determine if an
+                    // entrance can be placed on that side
+
+                    gRideEntranceExitPlaceStationIndex = trackElement->GetStationIndex();
+
+                    // get the ride entrance's side relative to the TrackElement
+                    Direction direction = (DirectionReverse(entranceExitCoords.direction) - tileElement->GetDirection()) & 3;
+                    const auto& ted = GetTrackElementDescriptor(trackElement->GetTrackType());
+                    if (ted.SequenceProperties[trackElement->GetSequenceIndex()] & (1 << direction))
+                    {
+                        // if that side of the TrackElement supports stations, the ride entrance is valid and faces away from
+                        // the station
+                        entranceExitCoords.direction = DirectionReverse(entranceExitCoords.direction);
+                        gRideEntranceExitPlaceDirection = entranceExitCoords.direction;
+                        return entranceExitCoords;
+                    }
+                } while (!(tileElement++)->IsLastForTile());
+            }
+        }
+        gRideEntranceExitPlaceDirection = INVALID_DIRECTION;
+        return entranceExitCoords;
+    }
+
 } // namespace OpenRCT2
