+++ src/openrct2-ui/windows/RideConstruction.cpp
@@ -7,6 +7,7 @@
  * OpenRCT2 is licensed under the GNU General Public License version 3.
  *****************************************************************************/
 
+#include "../interface/ViewportInteraction.h"
 #include "../ride/Construction.h"
 
 #include <limits>
